Add files via upload
Version of the script that only accepts files as the -s flag (.mol, .sdf and .smi/.smiles). 
If these files contain multiple molecules, it will loop through and perform the alignment on all of them. (Output will be 1 large SDF file containing all the molecules)
Added Tanimoto calculation in both commandprompt feedback aswell as in the SDF-file
Added error handling during the constrainedEmbed